+++ glean-core/tests/ping.rs
@@ -24,7 +24,7 @@ fn write_ping_to_disk() {
     });
     counter.add(&glean, 1);
 
-    assert!(ping.send(&glean).unwrap());
+    assert!(ping.submit(&glean).unwrap());
 
     assert_eq!(1, get_queued_pings(glean.get_data_path()).unwrap().len());
 }
@@ -45,7 +45,7 @@ fn disabling_upload_clears_pending_pings() {
     });
 
     counter.add(&glean, 1);
-    assert!(ping.send(&glean).unwrap());
+    assert!(ping.submit(&glean).unwrap());
     assert_eq!(1, get_queued_pings(glean.get_data_path()).unwrap().len());
     // At this point no deletion_request ping should exist
     // (that is: it's directory should not exist at all)
@@ -60,7 +60,7 @@ fn disabling_upload_clears_pending_pings() {
     assert_eq!(0, get_queued_pings(glean.get_data_path()).unwrap().len());
 
     counter.add(&glean, 1);
-    assert!(ping.send(&glean).unwrap());
+    assert!(ping.submit(&glean).unwrap());
     assert_eq!(1, get_queued_pings(glean.get_data_path()).unwrap().len());
 }
 
@@ -76,10 +76,10 @@ fn empty_pings_with_flag_are_sent() {
     // No data is stored in either of the custom pings
 
     // Sending this should succeed.
-    assert_eq!(true, ping1.send(&glean).unwrap());
+    assert_eq!(true, ping1.submit(&glean).unwrap());
     assert_eq!(1, get_queued_pings(glean.get_data_path()).unwrap().len());
 
     // Sending this should fail.
-    assert_eq!(false, ping2.send(&glean).unwrap());
+    assert_eq!(false, ping2.submit(&glean).unwrap());
     assert_eq!(1, get_queued_pings(glean.get_data_path()).unwrap().len());
 }